Added some new regression tests to do_mcmc_tests (catching bad input files, as in do_imfit_tests).
Added missing input files for config_file_parser.cpp unit tests to repo.

+++ core/mcmc_main.cpp
@@ -429,7 +429,7 @@ int main(int argc, char *argv[])
     if (paramLimits[i].fixed == 1)
       nFreeParams--;
   }
-  nDegFreedom = theModel->GetNValidPixels() - nFreeParams;
+  nDegFreedom = (long)(theModel->GetNValidPixels() - nFreeParams);
   printf("%d free parameters (%ld degrees of freedom)\n", nFreeParams, nDegFreedom);
 
   theModel->PrintDescription();
